requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.13.3
olefile==0.44
opencv-python==3.3.0.10
Pillow==4.3.0
pkg-resources==0.0.0
pytesseract==0.1.7
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
urllib3==1.22
Werkzeug==0.14.1
